requirements.txt
asgiref==3.2.10
autopep8==1.5.7
certifi==2021.5.30
charset-normalizer==2.0.3
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.1
django-admin-honeypot==1.1.0
django-admin-thumbnails==0.2.6
django-environ==0.4.5
django-import-export==2.5.0
django-session-timeout==0.1.0
et-xmlfile==1.1.0
idna==3.2
MarkupPy==1.14
mercadopago==2.0.7
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.3.0
psycopg2-binary==2.9.1
pycodestyle==2.7.0
python-decouple==3.4
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
tablib==3.0.0
toml==0.10.2
urllib3==1.26.6
xlrd==2.0.1
xlwt==1.3.0
